requirements.txt
numpy
opencv-python
hnswlib
imageio
torch
torchvision
pillow
tqdm
imagehash